decouple cli and server code
