+++ chaincash_cli/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "chaincash_cli"
+version = "0.1.0"
+edition = "2021"
+license = "CC0-1.0"
+
+[dependencies]
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
+tracing = "0.1.37"
+clap = { version = "4.3.23", features = ["derive"] }
+listenfd = "1.0.1"
+tokio = { version = "1", features = ["full"] }
+chaincash_server = { path = "../chaincash_server" }
